refactor: nfe group file names
