+++ src/dfe/nfe/schemas/groupB.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 import { zCustom } from "@/utils/zCustom";
 
-import { schemaNfeNfRef } from "./GroupBA";
+import { schemaNfeNfRef } from "./groupBA";
 
 export const zUfCode = () =>
   z.enum([
