+++ src/dfe/nfe/schemas/groupC.ts
